backend/requirements.txt
flask==3.0.0
mfrc522
RPi.GPIO